requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
grpcio==1.75.1
hf-xet==1.1.10
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
jaxtyping==0.3.3
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
mapanything @ git+https://github.com/facebookresearch/map-anything.git@44cbc5a4a3960cab50c449797afef200b88b94b7
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
minio==7.2.18
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python-headless==4.10.0.84
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pillow_heif==1.1.1
plyfile==1.1.2
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pycryptodome==3.23.0
pyparsing==3.2.5
python-box==7.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
rerun-sdk==0.24.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trimesh==4.8.3
triton==3.4.0
typing_extensions==4.15.0
uniception==0.1.4
urllib3==2.5.0
wadler_lindig==0.1.7
Werkzeug==3.1.3
